Challenge: Tasty Tomato
Bigger bite!
Before we make this program respond to mouse interaction, let's make a little change. See the bite in the tomato? It's tiny now, make it bigger!

Tip: the bite is being drawn as an ellipse. To make it bigger, you have to replace the width and height of the bite ellipse with numbers greater than 30.