Project: Ad design
In the text challenges, you showed that you know how to use text commands. Now, use that knowledge to make an advertisement for a real or made-up product!

Change the big headline to the name of your product.
Add at least 2 more text function calls advertising the product.
Change the text size so that you display text in at least 3 sizes.
Add a relevant drawing with at least 3 shape or image function calls.
Bonus: Play around with color and positioning to make it more flashy. If you want to go really wild, add an animation!