Project: Magic 8-Ball
Ever played with a magic 8-ball? It's a toy that displays a random message every time you shake it, such as "It is certain" or "My sources say no." In this project, you'll create your own magic 8-ball using if/else statements, so that you can use your program to answer your deepest questions about life.

Change the if() to check if the answer variable equals a certain number and change the text of the starter message.
Tip: Click restart until you see the message show up.
Add else if() statements to check for all the other possible values of the answer variable, and show a different message for each value.
Tip: For easy testing of specific values, you can add a line like var answer = 1 below the line that generates random numbers, then click restart to test it. Make sure you remove that debugging line before you send the project for evaluation.
Bonus: Give every message a different style, by changing the color, adding illustrations, or changing the text styles. There are lots of text related commands in the documentation to play with!
You can play with the numbers in floor(random(1, 5)) as long as you have at least 3 possible values for the answer variable, but don’t remove floor() or random(). The combination is the best way to create a reasonably even statistical distribution of values. To understand why, re-read the article on generating random numbers.