Project: Make it rain
To make an animation of rain, it's best if we use arrays to keep track of the drops and their different properties. Start with this simple code and build on it to make a cool rain animation.

Add more drops to the arrays. You can do that by using a for loop and random() function at the beginning of the program, or using an array modification method during animation. You could even add more drops in response to user interaction.
Make the rain continuous by changing the position of each drop when it reaches one side of the canvas, using a conditional or % operator.
Bonus: Make an array of colors, so that every drop is a different color, or an array of speeds.
Bonus: Make other things rain, like snowflakes (using more shape commands) or avatars (using the image commands). If you use multiple commands to draw whatever is raining down, consider creating a custom function with parameters instead.