Challenge: Picture Painter
This program uses an object to describe a paintbrush with x, y, and image properties. Change the initial x and y, and pick a different image that would be fun to paint with.

Tip: If you want to see all the images you can use, click on the string passed to getImage - the handy image picker will appear.